So the reason it took so much time is that I tried two separate methods. Out of these 2 methods, only the last worked.
First, I tried to load the model locally and have everyone download it, which is a bad idea. The second dumb idea I had while making this first version was the concept that making a request was bad. What was bad was requiring an API key.

So for my second version I started out trying to make a request to hugging face's API, but I was having trouble with requesting without a key. So I took a break on slack and remembered about ai.hackclub.com so I am using that.

Next devlog is with my cli.